requirements.txt
fastapi # Core web framework
uvicorn[standard] # ASGI server
python-dotenv # Environment variable management
groq # Language Model + Whisper STT
google-auth # Google Authentication
google-api-python-client # Google Calendar API
